Create new repo for building/testing Xyce
I wanted to see if I could create a single repo that could be cloned
and used directly to build Trilinos, build Xyce, and run the test
suite with as little user intervention as possible.

This is, of course, only a first cut, and at the moment has only been
set up for BSD.  Next step is to add an Ubuntu build.  But not today.