Add config for Ubuntu Linux
This commit adds configures for Trilinos and Xyce that work on Ubuntu.  They
will very likely work on other distros of Linux.

At the moment, my scripts all key off of what uname says the system is,
and are unable to distinguish between distros.  Therefore, some hand-hacking
may be necessary for users of distros other than Ubuntu.  But it is not
going to be very much.