Make Build_And_Test_Xyce always boostrap Xyce
It is not always necessary to rebootstrap Xyce (that is, regenerate
"configure") but it doesn't hurt to do so.  And it does hurt not to do
so when needed.

The old logic only ran bootstrap if configure did not exist.  This is
broken if any files have been removed or added to the build by the
Xyce team.