Add OS X/Homebrew support
This commit adds reconfigures for "Darwin" (how OS X identifies itself
in uname) that work with Homebrew.

The scripts are set up to be extensible for other package managers, but
I've only fleshed out homebrew (because that's all I have).  I have a
stub for macports, but it just prints out a plea for help and exits.