Fix text on "read" commands
The two build scripts pause for user input to make sure that the
choice of NUMPROCS=8 is appropriate for the user's machine.

"read" pauses for input and requires ENTER to proceed.  The text said
"Hit any other key" (than Control-C) to proceed.  That's wrong.

NUMPROCS should be a command-line parameter, perhaps with "-n" or
something.  That would be better than just pausing so the user can
interrupt and edit the script to change it.