Force relative RPATH into Linux builds
With this commit, Linux builds should wind up with an RPATH that
allows the binary to find its own shared libraries irrespective of
where the installed code might be moved.

Ideally, this should make it possible to tar up the install directory
and untar it elsewhere and have it still work.  I am not sure that's
true yet, though.