Add CFLAGS for all builds
Xyce has one directory of C code, in the ksparse directory.  Ksparse
is a non-default linear solver, and the lack of CFLAGS in all the Xyce
reconfigure scripts means this solver is getting no compiler
optimization.  If it were used, it would be slower than optimal.

This commit adds CFLAGS="-O3" to all reconfigure scripts.