Expand on update instructions
The instructions for updating the XyceBundle checkout and getting the
latest development version of Xyce are correct and work properly with
recent versions of git.

However, we've found at least one system (RHEL7) that has a version of
git so old that the submodule update is only partially done by
following the instructions here.  I've expanded the instructions so
that one can still use XyceBundle on those systems by using an extra
git command.