Get tags list for Xyce testing from script
As of this week, Xyce_Regression contains a script that probes an
executable for its capabilities and suggests a tags list to use for
the most rigorous testing of that executable by the test suite.

Using this script allows Xyce developers to assure that the full
capabilities of every binary run through the suite are tested, and
that the tags list used is appropriate for that binary.

We often get folks who compile binaries that have reduced capability,
and when that happens they often report failures that are solely the
result of the binary they built being tested with tests that use
features they've compiled out.