Make Stokhos a standard part of the build
The Xyce team wants to make Stokhos part of the default build soon,
and so I'm changing XyceBundle to build and test it.

This requires a change to all the Trilinos reconfigure scripts, and
because of the way XyceBundle is currently set up, this commit will
NOT force a rebuild of Trilinos --- but one is required.

You should rm -rf your XyceLibs subdirectory after pulling this
commit.  *THAT* will force a rebuild.