Update FreeBSD gfortran version
FreeBSD updated its gcc package to 10.x this week, and with it came
gfortran10.

This commit updates my BSD builds to use that fortran.  We still use
clang++ and clang as C++ and C compilers, so this change doesn't
matter for most things.  Really the only thing the gfortran is used
for is getting linkage information to couple correctly to blas and
lapack libraries --- there is no Fortran in Xyce.

gcc9 is still available, but this keeps XyceBundle current on that
system.