Change Trilinos and Xyce default options
The Xyce team will be changing the features included in the default
build soon, and this requires changing how Trilinos is built.

While these things are not default *yet*, it is relatively easy to
enable them explicitly right now, and I'm doing so in this repo.

Note that the change in the Trilinos configuration requires that one
blow away any existing XyceLibs directory first, because
"Build_And_Test_Xyce" will not rebuilt Trilinos if a directory exists
for the current system already --- even if it contains an unsuitable
build.  I otter fix that someday, but not now.  Just remember that if
the trilinos reconfigure has changed, you have to get rid of the older
install.