Allow modification of NUMPROC
The need to edit the scripts to build on machines that have fewer (or
many more) than 8 processors has started to annoy me, and I'm changing
them to allow it to be specified on the command line.

Build_And_Install_Trilinos and Build_And_Test_Xyce now take two
optional arguments, the number of processors and the location of the
top level of the bundle.

The first argument is the number of processors, the second is the
location.

This is in preparation to support Cygwin builds, and I only have a
2-processor Windows virtual machine.  My Windows VM "mounts" my Unix
home directory with my XyceBundle clone, so it is building out of a
shared directory rather than a separate clone.  Editing the script
just so it works on Windows and breaking it on my main machine wasn't
what I wanted.