Add Cygwin configs
This commit adds reconfigures for Trilinos and Xyce to be built on
Cygwin (after all TPLs listed in the building guide have been
installed, just as for all other systems).

The Trilinos config actually, definitely works.  The Xyce
config *should* work, but I haven't yet tested it.  I'm pushing so
that I can do a git clone of the XyceBundle repo on my Windows VM
instead of using a shared folder, which appears to slow *everything*
down badly.