Fix incorrect Trilinos cmake option
I had an option to enable LOCA that did nothing --- as it turns out,
LOCA gets enabled anyway, but the option I had in there was wrong
according to the building guide.

This seems to have no effect, but bringing it in line with the
building guide is the right thing to do.