Correct outdated error message
The "Build_And_Install_Trilinos" and "Build_And_Test_Xyce" scripts are
intended to be run while the current working directory is the top of
the XyceBundle clone, but they are set up to allow a user to decide to
run from elsewhere.

If invoked with no arguments, the scripts look in the current working
directory for Build_And_Install_Trilinos, and if it is found we are
safe to assume we're in the top of the XyceBundle tree.

But if it is not found, as would happen if the user did:
   git clone .../XyceBundle
   ./XyceBundle/Build_And_Install_Xyce

then we spit out an error message.  That error message should tell you
how to make it work, but it was outdated, and was unchanged when I
added the ability to specify the number of processors to be used by
the build and test scripts.

If one really wants to run these scripts from somewhere other than the
XyceBundle tree (e.g. to separate build directories and "install"
directories from the tree), one can easily do so by adding arguments.

  cd ~/someplace
  git clone .../XyceBundle.git
  cd ~/someplace_else
  ~/someplace/XyceBundle/Build_And_Test_Xyce 8 ~someplace/XyceBundle

The first argument is the number of cores, and the second argument is
where to find the top of the XyceBundle tree.  Build directories,
Trilinos install, and Xyce install prefix will be in ~/someplace_else,
leaving ~/someplace/XyceBundle pristine.