Add Test_Installed_Xyce script
This new script is invoked in the same manner as
"Build_And_Test_Xyce", and does almost the same thing --- only instead
of leaving the built version of Xyce in the build directory and
testing it from there, installs it in a subdirectory
Xyce_Installs/${UNAME}_Serial of the XyceBundle tree, and tests the
installed copy.

It will reuse the same build directory that was created by
Build_And_Test_Xyce, so if invoked right after that script it won't
actually do any building --- just installing and testing.

There are several tests that only work in installed versions, and
several other tests that only work from build directories, so doing
both makes sure that everything gets tested.

As of this commit, Xyce's "suggestXyceTagList.sh" script improperly
adds the tag to test verilog-a plugins if Xyce is built as a shared
build and with shared libraries, but does NOT test that ADMS is
installed --- so if you run Test_Installed_Xyce without ADMS being
installed the VERILOGPLUGIN tests will fail.

I am fixing that oversight upstream, but for now, either install ADMS
or expect those two tests to fail.