Update reconfigure for FreeBSD
Clang loves to complain when you specify an argument that is not used
by the compiler.  For example, if you specify "-Wl,..." (which means
"pass this argument to the linker") on a straight compile line, it'll
complain that it is a compiler not a linker and it is ignoring the
linker arguments.  This is a pointless warning --- we *KNOW* it's for
the linker, it is OK for everything but the linker to ignore it.

The Xyce autoconf process now forces a "-Wl,-rpath..." into the flags
automagically and so now Clang bitches constantly on every file.

This commit adds the magic incantation to tell clang to shut up and
ignore the unused arguments like a good boy.