Update FreeBSD fortran compiler
Xyce does not use fortran, but some packages in Trilinos do (packages
that Xyce doesn't even use).  This means that Xyce's build needs to
know what fortran compiler was used to compile Trilinos so it can link
to the libraries needed for Fortran.

Xyce on FreeBSD uses clang and clang++, but there is not yet a
fully functional Fortran compiler in the llvm project, so we have to
use gfortran in the Trilinos build, and have to tell Xyce which
compiler we used.

I was using gfortran10 when I created XyceBundle, but that was a long
time ago.  Let's use a newer one.