Add some more info to the README
I have upgraded my system to FreeBSD 13.2, and so now Xyce is known to
build and run there, too.  It's faster, by the way, because of the use
of Clang 16 instead of the Clang 10 that was used on FreeBSD 12.x.

I took this opportunity to update some documentation of XyceBundle.
Primarily this is adding more detail about what you can do after
you've built it and run the tests.  Like, say, where to find the
binary you just built, and how to install it elsewhere.