Update submodules for Release 7.9.0
The Xyce team released version 7.9.0 some time ago, but I haven't updated
XyceBundle to go with it because there are a number of bugs in both
Xyce and Trilinos 12.12.1 that make it crash on a number of tests on my
FreeBSD 13.4 system.  These have been reported to the Xyce team along with
their fixes, but none of those fixes are in Xyce 7.9.0, and the fixes
to Trilinos will never be put into the very ancient version 12.12.1.

Trilinos has been fixed on its main branch, though, but that branch
doesn't work with Xyce's autoconf very well.

Since I might be the only person in the world using this repo on FreeBSD,
I'm leaving well enough alone and just making this commit so that at least
this bundle should work on Linux.  I'm doing it from a Linux machine.

If a second person in the world turns out to be using this on FreeBSD, my
only option would be to fork trilinos and add the the small handful of
fixes to the 12.12.1 branch on my fork, then change the submodule on this
repo to point at my fork instead of the official repo.  I am loathe to do
that if nobody cares.

This is very likely my last commit to this XyceBundle repo, because it
appears that the Xyce team will soon (after many years of talking about it)
be replacing the autoconf build with a supported, complete CMake build.
Once they do that, I'm abandoning this metarepo, which exists entirely to
make using the autoconf system easier for users, and to assure that the
correct (ancient) version of Trilinos is being used.