Update submodules after root commit change
The Xyce team has updated their git procedures to facilitate pushing
code to github, and that involved restarting git history of the master
branch from a new root.  The old history is now in the dead-ended
'old_master' branch.

This commit simply updates the submodules so they point at the new
master branch.  There is no code change between this and release 7.9.